phiData/requirements.txt
phidata
duckduckgo-search
yfinance
groq
fastapi[standard]
sqlalchemy
python-dotenv
chromadb
openai
pypdf